fix: silence dead-code and signed/unsigned comparison warnings
Category 1 (dead code):
- Remove unused static fns: vmm dump_memory_map, tty tty_ring_empty,
  virtio-net virtio_fill_rx, terminal dbg; drop dead var pci page_base
  and tetris score.
- Inline debug-only locals into debug_* macros (fat/ahci/acpi/ioapic/trap).
- Annotate intentionally-disabled test fns (systest) and libk-only malloc
  helpers with __attribute__((unused)); mark unused params with it.

Category 2 (sign-compare in pmm.c):
- alloc_pages() number: int32_t -> uint64_t (drop number<0 guard); update
  pmm.h decl.
- subpage_pool.alloc_count: int -> uint32_t; free_subpage slot: int ->
  uint32_t (drop redundant slot>=0 guard).

+++ kernel/include/kernel/pmm.h
@@ -94,7 +94,7 @@ struct Zone
 };
 
 uint64_t page_init(struct Page * page, uint64_t flags);
-struct Page * alloc_pages(int32_t zone_select, int32_t number, uint64_t page_flags);
+struct Page * alloc_pages(int32_t zone_select, uint64_t number, uint64_t page_flags);
 void free_pages(struct Page * page,int32_t number);
 uint64_t page_clean(struct Page * page);
